cleanup: remove duplicated tests accross endpoints but using the same code
